requirements.txt
click==6.6
Flask==0.11.1
Flask-Script==2.0.5
Frozen-Flask==0.12
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
py-gfm==0.1.3
PyYAML==3.12
Werkzeug==0.11.10
